+++ NoFilter.cpp
@@ -3,8 +3,8 @@
 NoFilter::NoFilter() : Filter() {
 }
 
-HSB NoFilter::handleFilter(const unsigned long p_count,
-                           const unsigned long p_time,
+HSB NoFilter::handleFilter(const uint32_t p_count,
+                           const uint32_t p_time,
                            const HSB& p_hsb) {
     return p_hsb;
 }
